+++ src/components/TextForm.jsx
@@ -1,6 +1,5 @@
-import { useState } from "react";
 import styled from "styled-components";
-import { postQuestion } from "../services/postQuestion";
+import useTextForm from "../hooks/useTextForm";
 
 export const TextArea = styled.textarea`
   display: flex;
@@ -17,7 +16,6 @@ export const TextArea = styled.textarea`
     color: var(--gray40);
   }
   &:focus {
-    /* outline: none; */
     outline: 1px solid var(--brown40);
   }
   @media (max-width: 480px) {
@@ -43,43 +41,16 @@ export const SendingButton = styled.button`
   ${({ isValid }) =>
     isValid === true
       ? `background: var(--brown40)`
-      : `background: var(--brown30)`}
+      : `background: var(--brown30)`};
 `;
 
-const TextForm = ({
-  placeholder,
-  buttonText,
-  subject_id,
-  mode,
-  onClose,
-  setSend,
-}) => {
-  const [textValue, setTextValue] = useState("");
-  const [isValid, setIsValid] = useState(false);
-
-  const handleTextChange = (e) => {
-    // TextArea 상태 변화
-    const nextValue = e.target.value;
-    setTextValue(nextValue);
-    // 공백 or 빈 문자열인 경우
-    setIsValid(nextValue.trim() !== "");
-  };
-
-  const handleSubmit = async () => {
-    if (!isValid) return;
-    try {
-      if (mode === "question") {
-        await postQuestion({ subject_id, content: textValue });
-        console.log("질문 등록 완료!");
-        onClose();
-        setSend(true);
-      }
-      //답변 부분 (mode == "answer")
-    } catch (error) {
-      console.log("질문 등록 실패");
-      console.error(error);
-    }
-  };
+const TextForm = ({ placeholder, buttonText, id, mode, onClose, setSend }) => {
+  const { textValue, isValid, handleTextChange, handleSubmit } = useTextForm({
+    mode,
+    id,
+    onClose,
+    setSend,
+  });
 
   return (
     <>
